build: add Libs.private field in libnl pkg-config file
In order to support static linking, the libnl pkg-config file should
indicate in its Libs.private field the libraries that libnl-3.0.a
requires. The LIBS variable contains the appropriate list of
libraries: -lm in all cases, and -lpthread when pthread support is
enabled. This allows to statically link applications against libnl
properly.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
[Retrieved from:
https://git.buildroot.net/buildroot/tree/package/libnl/0001-build-add-Libs.private-field-in-libnl-pkg-config-fil.patch]
Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>