xfrm: remove superfluous xfrm_usersa_id from dump request
When being used with the dump flag, the kernel dispatches the request to
xfrm_dump_sa. This function does not expect the netlink message to
contain a struct xfrm_usersa_id as data payload of the netlink message.
Instead it interprets this payload as xfrm attributes. With the current
implementation, due to the sa_id being memset to zero, it causes the
kernel to complain about '24 bytes leftover after parsing attributes
[...]'.
This patch removes payload to get rid of the kernel complaint.

Signed-off-by: Thomas Egerer <thomas.egerer@secunet.com>

https://github.com/thom311/libnl/pull/233