lib: fix documentation of nl_cache_dump_filter to have @params optional
It doesn't make sense to call nl_cache_dump*() without dump options.
That is, because the dump options determine how the output works.

The @params parameter was thus wrongly specified as optional. Fix the
documentation. Probably this anyway lead to a crash, and hopefully
nobody was using that. However, be graceful and try to workaround it,
to avoid potentially crashing an application that previously worked.

https://github.com/thom311/libnl/issues/206