lib: allow to compare incomplete objects in nl_object_identical()
nl_object_identical() gets the ID attributes and then checks whether
these attributes are identical for both objects, by calling the compare
implementation.

If an object doesn't have all ID attributes set, previously there was no
way that this object would compare identical to anything.

That seems not right. For example, kernel has for many objects a rather
fuzzy understanding of what constitutes the ID attributes of an object.
For routes and routing rules for example, most attributes are part of
their ID, but also most of them are commonly missing and unset (note
that libnl3's understanding of about equality for routes is not correct
and does not correspond to kernels understanding. That is a known bug).

Anyway, such incomplete objects, if we have them at hand, we still shall
give them a chance to be evaluated as identical. Of course, and object
that has a certain ID attribute set must not compare identical to an
object that has the attribute missing. But if both objects miss the
attribute, they shall still be able to compare identical.

While at it, use uint64_t for the attribute mask. That corresponds to
the object's ce_mask and the API of oo_compare().

Note that if the object's ops would return id-attributes as 0xFFFFFFFF,
it previously would choose the intersection of the present attributes.
That is not correct either. If the ops don't specify any attributes,
then all present attributes shall be compared, not those that are
present in both.