route: remove incorrect nl_addr_valid() from rtnl_route_nh_set_newdst(), etc.
nl_addr_valid() expects an address family as argument, not the length.
It also expects an address in string form, not in binary.

Those checks were wrong. Also, it seems not necessary to validate the
argument, purely based on some criteria of the argument alone. Just set
the provided address.

Drop those checks.

This is relevant for NEWDST and ENCAP_MPLS DST processing.

Fixes: 760d74f99c88 ('route: Add support for MPLS encap')
Fixes: 0a6d27ce90a1 ('route: Add support for MPLS address family')

[thaller@redhat.com: split original patch and rewrite commit message].