route/mdb: hide rtnl_mdb_entry_alloc() from public API
"struct rtnl_mdb_entry" is a relatively simple object that keeps a few
data fields. But there are no public setters in the API, so it's not
useful to be able to allocate such an entry.

Hide from public API.

Also, it was already not listed in libnl-route-3.sym, and thus the
symbol was hidden anyway.