route: drop bitrate,sample-point getters/setters from can link
These accessor functions just get/set a field of ci_data_bittiming,
which can already be get/set as a whole. It seems redundant, so drop it.

It's really important that things are principle at all (like, setting
ci_data_bittiming). It's not clear that we need such convenience
functions, if you can achieve the same result otherwise.

Yes, we have similar accessors already. I think that is not a good
approach though.

If they turn out very useful, we can still revert this patch.