tests: add fixture/teardown for tests to run in separate netns
We run the unit tests as non-root user, so usually we wouldn't have
CAP_NET_ADMIN permissions to change networking. Also, we wouldn't want
that our unit tests depend on the networking of the test host (or
changes it).

For each test, enter a new network namespace (and user/mnt namespace).
There we will have the necessary permissions, and we are in full control
of the things in the namespace.

Note yet used.