clang-format: add ".clang-format" from linux kernel
Libnl historically is closely related to kernel development and follows
the same coding style. Copy kernel's format specification, in the hope to
be useful. There is no plan to reformat existing code, but with this
file in place, a developer can choose to selectively reformat the code.

Taken from linux kernel at [1]. The file is unmodified, except for the
"ForEachMacros".

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/.clang-format?id=56e337f2cf1326323844927a04e9dbce9a244835