route/mdb: add support for MAC multicast entries
The kernel bridge now support (permanent) forwarding of MAC multicast
using the MDB.  Internally the kernel use AF_UNSPEC, but we remap this
here to AF_LLC for the benefit for nl_addrs.

To test, put `nl-monitor mdb` in the background.  Then, with a bridge
and at least one port, run the following command:

    # nl-monitor mdb &
    # bridge mdb add dev br0 port eth0 grp 01:02:03:c0:ff:ee vid 1 permanent
    dev 9
    port 3 vid 1 proto 0x0000 address 01:02:03:c0:ff:ee

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>