netlink/private: add internal helper utils
inet_ntop() is documented to fail, so we have various places
with pointless (and wrong) error checking. Well, it can fail
if we pass an unexpected address family (which we must not and
assert against that), or if we pass an invalid string buffer (which we
must not, and cannot meaningfully assert for that). So it can only fail
in case of a bug and there is no need for error checking.

Yes, libc could theoretically fail, but if it fails on such a function that
requires no memory allocation, then it really needs to be fixed.