github: run unit tests several times and directly
`make check` swallows the output, which is kinda important
to debug a problem. I guess, I could print the log file, but
it seems simpler to just run the test directly.

While at it, run it a few times. It's fast enough, and with
the randomization, it might actually hit a faulty code path.