route: construct all-zero addresses for default route destination
A default route is equivalent to a 0.0.0.0/0 or ::/0 route, so we should
construct the dst as such with a all-zero address.

Since this breaks the assumption that a dst with a 0 address length is a
default route, switch to checking the prefix length being 0, and make
sure that there is an address part that is all-zero.

This ensures we will print the actual dst in case the address is not
zero, or does not exist.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>