bridge: don't normalize the u8 argument in rtnl_link_bridge_set_vlan_filtering() to boolean
In bridge-info, we have two u8 attributes (vlan-filtering and
stats-enabled). In both cases, kernel only expects there a boolean
value (either zero or one).

In case of vlan-filtering, I think kernel actually doesn't care, and
would treat any non-zero value as true. In case of stats-enabled, kernel
would reject values outside the values zero or one.

Previously, libnl3 would normalize the boolean value for vlan-filtering,
but not for stats-enabled. That is at least inconsistent, in particular
considering that kernel requires a normalized value for stats-enabled, but
not for vlan-filtering.

Our public API has uint8_t parameters (and not bool). That makes sense,
as it follows the netlink API.

It's not clear how to handle these boolean u8 attributes best. Should
the API be bool or uint8_t? Should we normalize the boolean values?

In any case, do something consistently. For no particular reason, choose
to not add additional logic. The user can set any value, whether that
value makes sense, it's up to them.