python: reformat all Python files with python-black
Automated code formatting is great. Use python-black for that ([1]).
Black us not configurable much (uncompromising) and follows PEP8.
It thus generates quite a standard, pleasing format.

But the point isn't whether the formatting is always the most
subjectively pleasing. The point is that the style is generated and
enforceable by an automated tool. It takes away places for nit picking,
while resulting in a more consistent code base.

As this reformats the entire code base, it introduces a large diff and
interferes with tools like git-blame. That will be somewhat alleviated
by adding a "blame.ignoreRevsFile" for git in the next commit.

The change is done automatically, by running python-black version
black-22.8.0-2.fc38.noarch:

    $ black .

The result is the same as

    $ ( git grep -l '#!.*\(PYTHON\|python\)' ; git ls-files '*.py' '*.py.in' ) \
        | sort -u \
        | xargs -d '\n' black

Note that the exact result depends on the version of black. Currently
that version is black==22.12.0 from pip. Yes, this means when we bump
the version in the future, some reformatting will happen.

[1] https://github.com/psf/black