route/link: improve handling of IFLA_INET6_CONF
On the netlink API, IFLA_INET6_CONF is a list of uint32_t integers, but
the number of actually provided values depends on the kernel version (and
the DEVCONF_MAX value that it was compiled with.

We clone the kernel header in our source tree, so DEVCONF_MAX from the
libnl3 build may differ from the running kernel. We need to keep track
how many values kernel actually provides.