clang-format: add "tools/clang-format.sh" script
Historically is libnl3 not formatted with clang-format, as the project
is old.

However, we have a mostly usable ".clang-format" file. Also, various new
files follow the automated formatting.

Add a script for helping with the reformatting.

Also, the script has an exclude list for files that should not be
formatted. That means, new files will be required to follow the
formatting. That is considered a benefit.