include: move "include/netlink-private/genl.h" to "lib/genl/nl-priv-genl.h"
This header is entirely private to the source files in "lib/genl".
It's confusing to keep it separate. Place it beside the source files,
which can use it.

I guess, I just disagree with this notion, that all headers must be
under "include/" directory. Not, if the header is entirely local to one
module.