include: move "include/netlink-private/route/utils.h" to nl-intern-route
This header has a very specific purpose. To give access to the unit
tests to some internals.

It only is usable to parts that statically link with libnl-route-3
sources (e.g. "tests/check-direct") and libnl-route-3 itself.

The point is that the symbol there is not exported by the libnl-route-3
shared library, so a test that dynamically links against libnl-route-3
couldn't access them. Hence the internal purpose of static linking the
test with libnl-route-3 sources, and a special header for that.

Move "include/netlink-private/route/utils.h" to a separate place, to
make that usage clearer.