idiag: "fix" license for "idiag-socket-details" tool
This tool had a GPL-2.0-only license text when it was added.
That is not what we require for libnl3 contributions.

Since being added, the file was only touched by me (Thomas Haller) and
Yegor Yefremov. Those are all trivial contributions.

This leaves Joe Damato, who initially contributed the file. Joe
agrees with this adjustment of the license (Thanks!!).

Note that there is also a line:

   Copyright (c) 2013 Sassano Systems LLC <joe@sassanosystems.com>

Apparently, the copyright for the contribution is also with Joe, so
there is no further problem.

Fixes: c97c8c2bfdb0 ('Add idiag-socket-details')

https://github.com/thom311/libnl/pull/354