license: fix/adjust license for "src/nl-cls-add.c"
In commit ef858fb492df ('- Reworked the classifier interface. - Added
initial ematch support - Added support for the basic classifier - Added
support for the cgroup classifier'), the license headers were added as
GPL-2.0-only. That is in conflict with the licences of the project and a
problem.

We would either have to delete the code, or "change" the license.

The file only had (relevant) contributions by Thomas Graf. There are
some trivial modifications by myself (Thomas Haller), and I agree with
this change. There is also a small change by another contributor, which
does seem trivial enough to not prevent changing the license:

- 4ff512b05456 [Cong Wang] ('correct --protocol option for nl-cls-* utilities')
- c52087601198 [Peter Wu] ('trivial: whitespace-only fixes for src and lib')
- cee0b1b50c2e [Yegor Yefremov] ('Add SPDX identifiers')

See `gitk src/nl-cls-add.c`.