build: add "-Wno-portability" to AC_INIT_AUTOMAKE()
Silence warnings with "-Wno-portability".

    Makefile.am:1130: warning: '%'-style pattern rules are a GNU make extension
    Makefile.am:1131: warning: dir $@: non-POSIX variable name
    Makefile.am:1131: (probably a GNU make extension)
    Makefile.am:1134: warning: '%'-style pattern rules are a GNU make extension
    Makefile.am:1137: warning: patsubst %,%.build-headers-test.o,$(public_headers: non-POSIX variable name
    Makefile.am:1137: (probably a GNU make extension)

We care about portability, but only to the extend that we test configurations
via CI. As all our current CI passes, the warning is bogus and not something we
will fix.

If you come up in a situation where this is a problem, then we have a real
scenario at hand and something to fix. Unless that happens, we target a real
implementation and not the make specification.